modified 104 fibonacci

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,7 +10,7 @@
 void printFibonacciNumbers(int n)
 {
 	unsigned long int x = 1, y = 2, z;
-	unsigned long int counter = 1;
+	int counter = 1;
 
 	while (counter <= n)
 	{
